requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.5.2
attr==0.3.2
attrs==22.1.0
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.5
dashboard==0.0.6
databases==0.6.1
Django==4.0.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-leaflet==0.28.3
django-phonenumber-field==7.0.0
django-queryset-csv==1.1.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
field==0.2.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
leaflet==0.0.3
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.4
orm==0.2.0.dev1
pandas==1.5.1
phonenumbers==8.12.57
Pillow==9.2.0
#psycopg2==2.9.4
psycopg2-binary==2.9.4
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.5
regex==2022.9.13
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
starlette==0.21.0
tqdm==4.64.1
typesystem==0.3.0.dev0
typing_extensions==4.4.0
tzdata==2022.5
unicodecsv==0.14.1
uvicorn==0.19.0
